# docs/swagger.yaml
definitions:
  domain.AuthResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/domain.User'
    type: object
  domain.GoogleLogin:
    properties:
      id_token:
        type: string
    required:
    - id_token
    type: object
  domain.Incident:
    properties:
      created_at:
        type: string
      description:
        type: string
      expires_at:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      lat:
        type: number
      lng:
        type: number
      reporter_id:
        type: string
      type:
        $ref: '#/definitions/domain.IncidentType'
    type: object
  domain.IncidentReport:
    properties:
      description:
        type: string
      lat:
        type: number
      lng:
        type: number
      type:
        $ref: '#/definitions/domain.IncidentType'
    required:
    - lat
    - lng
    - type
    type: object
  domain.IncidentType:
    enum:
    - accident
    - closure
    - hazard
    - traffic
    type: string
    x-enum-varnames:
    - IncidentAccident
    - IncidentClosure
    - IncidentHazard
    - IncidentTraffic
  domain.Place:
    properties:
      address:
        type: string
      category:
        type: string
      created_at:
        type: string
      external_id:
        type: string
      id:
        type: string
      lat:
        type: number
      lng:
        type: number
      name:
        type: string
      rating:
        type: number
    type: object
  domain.TravelHistory:
    properties:
      created_at:
        type: string
      id:
        type: integer
      lat:
        type: number
      lng:
        type: number
      speed_kph:
        type: number
      user_id:
        type: string
    type: object
  domain.User:
    properties:
      created_at:
        type: string
      display_name:
        type: string
      email:
        type: string
      id:
        type: string
      updated_at:
        type: string
    type: object
  domain.UserLogin:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  domain.UserRegistration:
    properties:
      display_name:
        type: string
      email:
        type: string
      password:
        minLength: 8
        type: string
    required:
    - display_name
    - email
    - password
    type: object
info:
  contact: {}
paths:
  /incidents:
    get:
      parameters:
      - description: Latitude
        in: query
        name: lat
        required: true
        type: number
      - description: Longitude
        in: query
        name: lng
        required: true
        type: number
      - description: Radius in meters
        in: query
        name: radius
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Incident'
            type: array
      summary: Get nearby incidents
      tags:
      - incidents
    post:
      consumes:
      - application/json
      parameters:
      - description: Incident details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.IncidentReport'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Incident'
      security:
      - ApiKeyAuth: []
      summary: Report a traffic incident
      tags:
      - incidents
  /places/{id}:
    get:
      parameters:
      - description: Place ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Place'
      summary: Get place details
      tags:
      - places
  /places/{id}/favorite:
    post:
      parameters:
      - description: Place UUID
        in: path
        name: id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Add a place to favorites
      tags:
      - places
  /places/favorites:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Place'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Get user's favorite places
      tags:
      - places
  /places/search:
    get:
      parameters:
      - description: Search text
        in: query
        name: text
        required: true
        type: string
      - description: Latitude
        in: query
        name: lat
        type: number
      - description: Longitude
        in: query
        name: lng
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Place'
            type: array
      summary: Search for places
      tags:
      - places
  /timeline:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.TravelHistory'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Get user travel history
      tags:
      - timeline
    post:
      consumes:
      - application/json
      parameters:
      - description: Location details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.TravelHistory'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Record user location
      tags:
      - timeline
  /users/google-login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Google ID Token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.GoogleLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.AuthResponse'
      summary: Login with Google
      tags:
      - users
  /users/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.UserLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.AuthResponse'
      summary: Login user
      tags:
      - users
  /users/profile:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
      security:
      - ApiKeyAuth: []
      summary: Get user profile
      tags:
      - users
  /users/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: User details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.UserRegistration'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.AuthResponse'
      summary: Register a new user
      tags:
      - users
swagger: "2.0"